refactored test objects
